+++ django/applications/catmaid/static/js/tools/tracing-tool.js
@@ -1139,6 +1139,36 @@
     }));
 
     this.addAction(new CATMAID.Action({
+      helpText: "Bookmark the active node or current location",
+      keyShortcuts: { ';': [ 186 ] },
+      run: function (e) {
+          var dialog = new CATMAID.BookmarkDialog(CATMAID.BookmarkDialog.MODES.MARK);
+          dialog.show();
+          return true;
+      }
+    }));
+
+    this.addAction(new CATMAID.Action({
+      helpText: "Go to a bookmarked skeleton",
+      keyShortcuts: { '\'': [ 222 ] },
+      run: function (e) {
+          var dialog = new CATMAID.BookmarkDialog(CATMAID.BookmarkDialog.MODES.SKELETON);
+          dialog.show();
+          return true;
+      }
+    }));
+
+    this.addAction(new CATMAID.Action({
+      helpText: "Go to a bookmarked node",
+      keyShortcuts: { '`': [ 192 ] },
+      run: function (e) {
+          var dialog = new CATMAID.BookmarkDialog(CATMAID.BookmarkDialog.MODES.NODE);
+          dialog.show();
+          return true;
+      }
+    }));
+
+    this.addAction(new CATMAID.Action({
       helpText: "Toggle display of skeletons in visibility group 1 (<kbd>Shift</kbd>: visibility group 2)",
       keyShortcuts: { 'HOME': [ 36 ] },
       run: function (e) {